refact: aplicar responsabilidad unica reprogramar
