+++ src/templates/flujos/reprogramarCita/index.ts
@@ -1,5 +1,9 @@
 export * from './step1Reprogramar';
-export * from './step3Reprogramar';
+export * from './step5Reprogramar';
 export * from './stepConfirmaReprogramar';
 export * from './noConfirmaReprogramar';
 export * from './seleccionaCitaReprogramar';
+export * from './step6Reprogramar';
+export * from './step7Reprogramar';
+export * from './stepHoraSeleccionada';
+export * from './stepSeleccionaFechaReprogramar';
